backend/pyproject.toml
[project]
name = "saferoute-app"
version = "0.1.0"
description = "LLM routing gateway with analytics and key management"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt<4",
    "cryptography>=48.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "openai>=2.34.0",
    "pandas>=3.0.2",
    "passlib[bcrypt]>=1.7.4",
    "posthog>=7.14.0",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "scikit-learn==1.6.1",
    "sentry-sdk",
    "sqlalchemy[asyncio]>=2.0.49",
    "textstat>=0.7.13",
    "uvicorn[standard]>=0.46.0",
    "xgboost>=3.2.0",
]
